transfer navBar item into separate component
